# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AirboxTechnical",
    "version": "1.0"
  },
  "paths": {
    "/User": {
      "post": {
        "tags": [
          "User"
        ],
        "operationId": "CreateUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AirboxTechnical.Models.CreateUserRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AirboxTechnical.Models.CreateUserRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AirboxTechnical.Models.CreateUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AirboxTechnical.Models.User"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AirboxTechnical.Models.User"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AirboxTechnical.Models.User"
                }
              }
            }
          }
        }
      }
    },
    "/UserLocation": {
      "post": {
        "tags": [
          "UserLocation"
        ],
        "operationId": "AddLocation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AirboxTechnical.Models.CreateUserLocationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AirboxTechnical.Models.CreateUserLocationRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AirboxTechnical.Models.CreateUserLocationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AirboxTechnical.Data.Models.UserLocation"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AirboxTechnical.Data.Models.UserLocation"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AirboxTechnical.Data.Models.UserLocation"
                }
              }
            }
          }
        }
      }
    },
    "/UserLocation/recent/{userId}": {
      "get": {
        "tags": [
          "UserLocation"
        ],
        "operationId": "GetLastLocation",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AirboxTechnical.Data.Models.UserLocation"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AirboxTechnical.Data.Models.UserLocation"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AirboxTechnical.Data.Models.UserLocation"
                }
              }
            }
          }
        }
      }
    },
    "/UserLocation/recent": {
      "get": {
        "tags": [
          "UserLocation"
        ],
        "operationId": "GetLastLocations",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AirboxTechnical.Data.Models.UserLocation"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AirboxTechnical.Data.Models.UserLocation"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AirboxTechnical.Data.Models.UserLocation"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/UserLocation/history/{userId}": {
      "get": {
        "tags": [
          "UserLocation"
        ],
        "operationId": "GetLocations",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AirboxTechnical.Data.Models.UserLocation"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AirboxTechnical.Data.Models.UserLocation"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AirboxTechnical.Data.Models.UserLocation"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AirboxTechnical.Data.Models.User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AirboxTechnical.Data.Models.UserLocation": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AirboxTechnical.Data.Models.User"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "AirboxTechnical.Models.CreateUserLocationRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "AirboxTechnical.Models.CreateUserRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AirboxTechnical.Models.User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}